+++ src/View/Components/Editor.php
@@ -2,6 +2,7 @@
 
 namespace Darvis\FluxFilemanager\View\Components;
 
+use Darvis\FluxFilemanager\Support\FluxFilemanagerConfig;
 use Illuminate\Contracts\View\View;
 use Illuminate\View\Component;
 
@@ -21,14 +22,14 @@ class Editor extends Component
     public function jsConfig(): array
     {
         return [
-            'filemanager_url' => config('flux-filemanager.url', '/filemanager'),
-            'popup_width' => config('flux-filemanager.popup.width', 900),
-            'popup_height' => config('flux-filemanager.popup.height', 600),
-            'resize_presets' => config('flux-filemanager.resize_presets', ['25%', '50%', '75%', '100%']),
-            'custom_width_min' => config('flux-filemanager.custom_width.min', 1),
-            'custom_width_max' => config('flux-filemanager.custom_width.max', 100),
-            'popup_blocked_message' => config('flux-filemanager.messages.popup_blocked'),
-            'filemanager_error_message' => config('flux-filemanager.messages.filemanager_not_found'),
+            'filemanager_url' => FluxFilemanagerConfig::url(),
+            'popup_width' => FluxFilemanagerConfig::popupWidth(),
+            'popup_height' => FluxFilemanagerConfig::popupHeight(),
+            'resize_presets' => FluxFilemanagerConfig::resizePresets(),
+            'custom_width_min' => FluxFilemanagerConfig::customWidthMin(),
+            'custom_width_max' => FluxFilemanagerConfig::customWidthMax(),
+            'popup_blocked_message' => FluxFilemanagerConfig::popupBlockedMessage(),
+            'filemanager_error_message' => FluxFilemanagerConfig::filemanagerNotFoundMessage(),
             'i18n' => trans('flux-filemanager::filemanager'),
         ];
     }